% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{calculate_ICI,PathObject-method}
\alias{calculate_ICI,PathObject-method}
\title{Calculate ICI index encapsulates the functions in the ConsinusClusterPlus package}
\usage{
\S4method{calculate_ICI}{PathObject}(obj, out_dir = "test/ConsensusCluster", ...)
}
\arguments{
\item{obj}{PathObject}

\item{...}{custom parameters}
}
\value{
PathObject
}
\description{
Calculate ICI index encapsulates the functions in the ConsinusClusterPlus package
}
