% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{ConsensusCluster,PathObject-method}
\alias{ConsensusCluster,PathObject-method}
\title{ConsinusCluster encapsulates the functions in the ConsinusClusterPlus package}
\usage{
\S4method{ConsensusCluster}{PathObject}(
  obj,
  maxK = 9,
  repcount = 50,
  pSample = 0.8,
  weightItems = NULL,
  pRow = 1,
  weightFeatures = NULL,
  distance = "euclidean",
  clusterAlg = "km",
  corUse = NULL,
  innerLinkage = "average",
  seed = 123456,
  out_dir = "test/ConsensusCluster"
)
}
\arguments{
\item{obj}{PathObject}

\item{all}{All parameters are the same as those of the R package ConsensusClusterPlus function}
}
\value{
PathObject
}
\description{
ConsinusCluster encapsulates the functions in the ConsinusClusterPlus package
}
