% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_S4_object.R
\name{CreatePathObject}
\alias{CreatePathObject}
\title{Create a PathObj S4 Object}
\usage{
CreatePathObject(
  gmt_path = NULL,
  gene_exp_counts = NULL,
  gene_map = NULL,
  gene_exp_TPM = NULL,
  ppi_data_path = NULL,
  cor_method = 1,
  mut_exp = NULL,
  count_log = TRUE,
  TPM_log = FALSE,
  GeneSetCollect = FALSE,
  sep = "\\t"
)
}
\arguments{
\item{gmt_path}{a gmt type file path It is recommended to use the 50Hallmark gene set}

\item{gene_exp_counts}{a data.frame or matrix gene exprssion or path the first column should be geneid}

\item{gene_map}{a gtf type file or data.frame with 5 columns included gene length start end chrid and geneid}

\item{gene_exp_TPM}{same as gene_exp_counts}

\item{ppi_data_path}{two columns node A and node B}

\item{cor_method}{1 pearson 2 spearman 3 kendall}

\item{mut_exp}{MAF or binary matrix}

\item{count_log}{if log2 Transform}

\item{TPM_log}{if log2 Transform}

\item{GeneSetCollect}{gmt transform to GeneSetCollect}

\item{sep}{Tabs}
}
\value{
PathObj S4 object
}
\description{
Create a PathObj S4 Object
}
