% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{DegreePathmutscore,PathObject-method}
\alias{DegreePathmutscore,PathObject-method}
\title{Calculate Individual Weight Hallmark Gene Set mutation burden for every sample}
\usage{
\S4method{DegreePathmutscore}{PathObject}(obj, ...)
}
\arguments{
\item{obj}{PathObj object}

\item{...}{custom parameters}
}
\value{
PathObj
}
\description{
Calculate Individual Weight Hallmark Gene Set mutation burden for every sample
}
