% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.network_summary}
\alias{.network_summary}
\title{Calculate the network properties of each Hallmark Gene Set network only in windows}
\usage{
.network_summary(object, nthreads = 4)
}
\arguments{
\item{object}{PathObj}

\item{nthreads}{Number of threads}
}
\description{
Calculate the network properties of each Hallmark Gene Set network only in windows
}
