% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.Plaw_pvalue}
\alias{.Plaw_pvalue}
\title{Functions for calculating power law distribution are not open to the public}
\usage{
.Plaw_pvalue(x, nthreads)
}
\arguments{
\item{x}{Degree}

\item{nthreads}{number of nthreads}
}
\description{
Functions for calculating power law distribution are not open to the public
}
