% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{Hclust,PathObject-method}
\alias{Hclust,PathObject-method}
\title{This function is hierarchical clustering Mainly depends on pvlcust}
\usage{
\S4method{Hclust}{PathObject}(obj, method.hclust = "complete", method.dist = "euclidean", nboot = 100)
}
\arguments{
\item{obj}{PathObject}

\item{method.hclust}{same in pvlcust function}

\item{method.dist}{same in pvlcust function}

\item{nboot}{same in pvlcust function}
}
\value{
PathObject
}
\description{
This function is hierarchical clustering Mainly depends on pvlcust
}
