% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{mut_enrich,PathObject-method}
\alias{mut_enrich,PathObject-method}
\title{Enrichment of somatic mutation in each Hallmark Gene Set or parhway network}
\usage{
\S4method{mut_enrich}{PathObject}(obj, my_pal, mut_classifcation, degree = NULL)
}
\arguments{
\item{obj}{PathObject}

\item{my_pal}{Color table in coding and non-coding}

\item{mut_classifcation}{The somatic mutation classification data table of the cohort
is composed of two columns, the first column is the gene name, and the second column
is the mutation division of the gene}

\item{degree}{Number of genes connected to somatic mutant genes in Hallmark Gene Set Network}
}
\description{
Enrichment of somatic mutation in each Hallmark Gene Set or parhway network
}
