% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{NetFeatureCalculate,PathObject-method}
\alias{NetFeatureCalculate,PathObject-method}
\title{Calculate Network Features of Hallmark Gene Set or Pathways for current cohort}
\usage{
\S4method{NetFeatureCalculate}{PathObject}(obj, ...)
}
\arguments{
\item{obj}{PathObj object}

\item{...}{custom parameters}
}
\value{
new PathObj
}
\description{
Calculate Network Features of Hallmark Gene Set or Pathways for current cohort
}
