% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{plot_mut_distribution,PathObject-method}
\alias{plot_mut_distribution,PathObject-method}
\title{This function is plot the distributation of muation in or out of Hallmark Gene Set or parhway network}
\usage{
\S4method{plot_mut_distribution}{PathObject}(obj, my_pal, mut_classifcation)
}
\arguments{
\item{obj}{PathObject}

\item{my_pal}{Color table in coding and non-coding}

\item{mut_classifcation}{The somatic mutation classification data table of the cohort
is composed of two columns, the first column is the gene name, and the second column
is the mutation division of the gene}
}
\description{
This function is plot the distributation of muation in or out of Hallmark Gene Set or parhway network
}
