% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{plot_point_fetsig,PathObject-method}
\alias{plot_point_fetsig,PathObject-method}
\title{Describe the signaficance of  characteristics of Hallmark Gene Set or parhway network}
\usage{
\S4method{plot_point_fetsig}{PathObject}(obj, my_pal)
}
\arguments{
\item{obj}{PathObject}

\item{my_pal}{color table}
}
\description{
Describe the signaficance of  characteristics of Hallmark Gene Set or parhway network
}
