% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_dispatch.R
\name{SelectBestClusterNum,PathObject-method}
\alias{SelectBestClusterNum,PathObject-method}
\title{SelectBestClusterNum for cluster}
\usage{
\S4method{SelectBestClusterNum}{PathObject}(obj, method = "hclust", k = NULL, ...)
}
\arguments{
\item{obj}{PathObject}

\item{method}{"hclust" or "Consensus"}

\item{k}{the number of cluster you want}
}
\value{
PathObject
}
\description{
SelectBestClusterNum for cluster
}
